changes routes rendering to another file
