+++ client/src/App.tsx
@@ -10,14 +10,9 @@ import { Navbar } from './components/Navbar/Navbar';
 import { Route, Switch } from 'react-router-dom';
 import { getRoutesAccordingToAccountType } from './pages/routes/route';
 import NotFound from './pages/NotFound/NotFound';
+import RenderRoutes from './pages/routes/routes';
 
 function App(): JSX.Element {
-  const { profile } = useAuth();
-  let accountType;
-
-  if (profile) accountType = profile.accountType;
-
-  const routes = getRoutesAccordingToAccountType(accountType);
   return (
     <ThemeProvider theme={theme}>
       <SnackBarProvider>
@@ -25,14 +20,7 @@ function App(): JSX.Element {
           <SocketProvider>
             <CssBaseline />
             <Navbar />
-            <Switch>
-              {routes.map((item, i) => (
-                <Route key={i} path={item.resource} component={item.component} />
-              ))}
-              <Route path="*">
-                <NotFound />
-              </Route>
-            </Switch>
+            <RenderRoutes />
           </SocketProvider>
         </AuthProvider>
       </SnackBarProvider>
